clippy and rearrange Regexp and RegexpType trait impls
